.history/requirements_20250206231844.txt
fastapi
uvicorn
pytesseract
pdf2image
python-multipart
python-dotenv   # Přidej tuhle řádku
